new version of xterm-addon-attach doesn't cause build errors
