+++ SingularityUI/app/components/logs/less/TaskLessButton.jsx
@@ -5,8 +5,7 @@ import OverlayTrigger from 'react-bootstrap/lib/OverlayTrigger';
 import ToolTip from 'react-bootstrap/lib/Tooltip';
 
 import SimplestModal from '../../common/modal/SimplestModal';
-
-import TaskLessTerminal from './TaskLessTerminal';
+import LessTerminal from './LessTerminal';
 
 export default class TaskLessButton extends Component {
 
@@ -44,9 +43,7 @@ export default class TaskLessButton extends Component {
         </OverlayTrigger>
         <SimplestModal ref="modal" bsSize="lg">
           <Modal.Body style={{ padding: 0 }}>
-            <TaskLessTerminal
-              host={this.props.host}
-              port={config.lessTerminalPort}
+            <LessTerminal
               task={this.props.task}
               path={this.props.file.fullPath}
               onClose={() => { this.refs.modal.hide() }}
