+++ SingularityUI/package.json
@@ -64,6 +64,7 @@
     "underscore": "~1.8.0",
     "uuid": "^3.0.0",
     "xterm": "^4.3.0",
+    "xterm-addon-attach": "^0.5.0",
     "xterm-addon-fit": "^0.3.0"
   },
   "license": "Apache-2.0",
